refactor/libretroize file and directory code
