+++ src/emu/crsshair.c
@@ -16,7 +16,7 @@
 #include "xmlfile.h"
 #include "crsshair.h"
 
-
+#include "retromain.h"
 
 /***************************************************************************
     CONSTANTS
@@ -161,17 +161,17 @@ static void create_bitmap(running_machine *machine, int player)
 	{
 		/* look for user specified file */
 		sprintf(filename, "%s.png", global.name[player]);
-		global.bitmap[player] = render_load_png(OPTION_CROSSHAIRPATH, NULL, filename, NULL, NULL);
+		global.bitmap[player] = render_load_png(crosshairpath, NULL, filename, NULL, NULL);
 	}
 	else
 	{
 		/* look for default cross?.png in crsshair\game dir */
 		sprintf(filename, "cross%d.png", player + 1);
-		global.bitmap[player] = render_load_png(OPTION_CROSSHAIRPATH, machine->gamedrv->name, filename, NULL, NULL);
+		global.bitmap[player] = render_load_png(crosshairpath, machine->gamedrv->name, filename, NULL, NULL);
 
 		/* look for default cross?.png in crsshair dir */
 		if (global.bitmap[player] == NULL)
-			global.bitmap[player] = render_load_png(OPTION_CROSSHAIRPATH, NULL, filename, NULL, NULL);
+			global.bitmap[player] = render_load_png(crosshairpath, NULL, filename, NULL, NULL);
 	}
 
 	/* if that didn't work, use the built-in one */
